Task update

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -2,22 +2,26 @@
 /**
  * main - Entry point of the program
  *
- * Return: 0 on success
+ * Description: Calculates and prints the first
+ * 98 Fibonacci numbers
+ *  separated by commas and a space.
+ *
+ * Return: Always 0 (Success)
  */
 int main(void)
 {
 int count = 0;
-unsigned long int fib1 = 1;
-unsigned long int fib2 = 2;
-unsigned long int fib_next;
+unsigned long long fib1 = 1;
+unsigned long long fib2 = 2;
+unsigned long long fib_next;
 
-printf("%lu, %lu", fib1, fib2);
+printf("%llu, %llu", fib1, fib2);
 count += 2;
 
 while (count < 98)
 {
 fib_next = fib1 + fib2;
-printf(", %lu", fib_next);
+printf(", %llu", fib_next);
 fib1 = fib2;
 fib2 = fib_next;
 count++;
